Options folder to follow component standards
