+++ cmd/kube-proxy/app/init_windows.go
@@ -20,8 +20,6 @@ package app
 
 import (
 	"k8s.io/kubernetes/pkg/windows/service"
-
-	"github.com/spf13/pflag"
 )
 
 const (
@@ -33,11 +31,4 @@ func initForOS(windowsService bool) error {
 		return service.InitService(serviceName)
 	}
 	return nil
-}
-
-func (o *Options) addOSFlags(fs *pflag.FlagSet) {
-	fs.BoolVar(&o.WindowsService, "windows-service", o.WindowsService, "Enable Windows Service Control Manager API integration")
-	fs.StringVar(&o.config.Winkernel.SourceVip, "source-vip", o.config.Winkernel.SourceVip, "The IP address of the source VIP for non-DSR.")
-	fs.StringVar(&o.config.Winkernel.NetworkName, "network-name", o.config.Winkernel.NetworkName, "The name of the cluster network.")
-	fs.BoolVar(&o.config.Winkernel.EnableDSR, "enable-dsr", o.config.Winkernel.EnableDSR, "If true make kube-proxy apply DSR policies for service VIP")
 }
