+++ cmd/kube-proxy/app/options/osflags_others.go
@@ -1,7 +1,7 @@
 // +build !windows
 
 /*
-Copyright 2018 The Kubernetes Authors.
+Copyright 2021 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -16,8 +16,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package app
+package options
 
-func initForOS(service bool) error {
-	return nil
+import (
+	"github.com/spf13/pflag"
+)
+
+func (o *Options) addOSFlags(fs *pflag.FlagSet) {
 }
